cli: default call/download/up/shell to have core APIs
We can actually do this in the short term without breaking the current
UX. The modification needed was to change the OnSelectObjectLeaf
callback to run once the last command is being executed (i.e. we're
about to construct and submit the actual API query).

This allows the existing behavior to be retained (for now) while also
supporting continued chaining off of core object types like
Container/Directory/File too.

As part of this I also split the existing `dagger call` integ tests to
their own file, backfilled some coverage around the current behavior
when Container/Directory/File is the final return type and also added
new coverage for the new core chaining abilities.
