+++ cmd/dagger/call.go
@@ -6,21 +6,10 @@ import (
 	"github.com/spf13/cobra"
 )
 
-var (
-	experimentalIncludeCore bool
-)
-
 var callCmd = &FuncCommand{
 	Name:  "call",
 	Short: "Call a module function",
 	Long:  "Call a module function and print the result.\n\nOn a container, the stdout will be returned. On a directory, the list of entries, and on a file, its contents.",
-	Init: func(cmd *cobra.Command) {
-		cmd.PersistentFlags().BoolVarP(&experimentalIncludeCore, "experimental-include-core", "x", false, "Enable experimental inclusion of core APIs in the call command")
-	},
-	BeforeLoad: func(fc *FuncCommand) error {
-		fc.IncludeCore = experimentalIncludeCore
-		return nil
-	},
 	OnSelectObjectLeaf: func(c *FuncCommand, name string) error {
 		switch name {
 		case Container:
@@ -32,8 +21,7 @@ var callCmd = &FuncCommand{
 		case File:
 			c.Select("contents")
 		default:
-			// TODO: Check if it's a core object and sub-select `id` by default.
-			return fmt.Errorf("return type not supported: %s", name)
+			return fmt.Errorf("return type %q requires a sub-command", name)
 		}
 		return nil
 	},
